last commit

+++ app.py
@@ -1,5 +1,12 @@
-
 from flask import Flask,request,render_template
+
+
+
+
+
+
+
+
 import pickle
 import numpy as np
 import pandas as pd
@@ -15,15 +22,16 @@ def home():
 
 
 @app.route('/',methods=['POST'])
+
 def predict():
 
-    funding_rounds=int(request.form["funding_rounds"])
-    funding_total_usd=int(request.form['funding_total_usd'])
-    milestones=int(request.form['milestones'])
-    relationships=int(request.form['relationships'])
+    funding_rounds=float(request.form["funding_rounds"])
+    funding_total_usd=float(request.form['funding_total_usd'])
+    milestones=float(request.form['milestones'])
+    relationships=float(request.form['relationships'])
     lat=float(request.form['lat'])
     lng=float(request.form['lng'])
-    active_days=int(request.form['active_days'])
+    active_days=float(request.form['active_days'])
     
     data=['funding_rounds','funding_total_usd','milestones','relationships','lat','lng','active_days']
